Move HashMapDb to its own bundle
HashMapDb is a simple, in memory implementation of ILispDAO, the
interface that defines the API for the interaction with the Map-Server's
mappings database. Until now, HashMapDb was part of
mappingservice.implementation but with this commit it is moved to its
own mappingservice.inmemorydb bundle.

An important feature is that a ILispDAO implementation is now injected
into mappingservice.implementation at runtime through config-subsystem.
As a result, InMemoryDb can be easily swapped with other implementations.
