+++ artifacts/pom.xml
@@ -39,6 +39,11 @@ and is available at http://www.eclipse.org/legal/epl-v10.html
       </dependency>
       <dependency>
         <groupId>${project.groupId}</groupId>
+        <artifactId>mappingservice.inmemorydb</artifactId>
+        <version>${project.version}</version>
+      </dependency>
+      <dependency>
+        <groupId>${project.groupId}</groupId>
         <artifactId>mappingservice.netconf</artifactId>
         <version>${project.version}</version>
       </dependency>
