+++ commons/parent/pom.xml
@@ -231,7 +231,11 @@
         <artifactId>mappingservice.implementation</artifactId>
         <version>${lispflowmapping.version}</version>
       </dependency>
-
+      <dependency>
+        <groupId>org.opendaylight.lispflowmapping</groupId>
+        <artifactId>mappingservice.inmemorydb</artifactId>
+        <version>${lispflowmapping.version}</version>
+      </dependency>
       <dependency>
         <groupId>org.opendaylight.lispflowmapping</groupId>
         <artifactId>mappingservice.netconf</artifactId>
