+++ features/src/main/resources/features.xml
@@ -37,6 +37,7 @@
         <!-- <feature version="${clustering.services.version}">odl-adsal-clustering</feature> -->
         <feature version='${yangtools.version}'>odl-yangtools-binding</feature>
         <feature version='${yangtools.version}'>odl-yangtools-models</feature>
+        <feature version="${lispflowmapping.version}">odl-lispflowmapping-inmemorydb</feature>
         <bundle>mvn:com.google.guava/guava/${guava.version}</bundle>
         <bundle>mvn:org.opendaylight.lispflowmapping/mappingservice.yangmodel/${lispflowmapping.version}</bundle>
         <bundle>mvn:org.opendaylight.lispflowmapping/mappingservice.api/${lispflowmapping.version}</bundle>
@@ -45,6 +46,19 @@
         <!-- <bundle>mvn:org.opendaylight.lispflowmapping/mappingservice.clusterdao/${lispflowmapping.version}</bundle>  -->
         <configfile finalname="${configfile.directory}/lfm-mappingservice.xml">
             mvn:org.opendaylight.lispflowmapping/mappingservice.implementation/${lispflowmapping.version}/xml/config
+        </configfile>
+    </feature>
+
+    <feature name="odl-lispflowmapping-inmemorydb" description="OpenDaylight :: LISP Flow Mapping :: In Memory DB" version='${lispflowmapping.version}'>
+        <feature version="${mdsal.version}">odl-mdsal-broker</feature>
+        <feature version='${yangtools.version}'>odl-yangtools-binding</feature>
+        <feature version='${yangtools.version}'>odl-yangtools-models</feature>
+        <bundle>mvn:com.google.guava/guava/${guava.version}</bundle>
+        <bundle>mvn:org.opendaylight.lispflowmapping/mappingservice.yangmodel/${lispflowmapping.version}</bundle>
+        <bundle>mvn:org.opendaylight.lispflowmapping/mappingservice.api/${lispflowmapping.version}</bundle>
+        <bundle>mvn:org.opendaylight.lispflowmapping/mappingservice.inmemorydb/${lispflowmapping.version}</bundle>
+        <configfile finalname="${configfile.directory}/lfm-inmemorydb.xml">
+            mvn:org.opendaylight.lispflowmapping/mappingservice.inmemorydb/${lispflowmapping.version}/xml/config
         </configfile>
     </feature>
 
@@ -67,4 +81,5 @@
         <feature version="${lispflowmapping.version}">odl-lispflowmapping-mappingservice</feature>
         <bundle>mvn:org.opendaylight.lispflowmapping/mappingservice.neutron/${lispflowmapping.version}</bundle>
     </feature>
+
 </features>
