+++ mappingservice/implementation/pom.xml
@@ -8,9 +8,7 @@
     <version>1.3.0-SNAPSHOT</version>
   </parent>
 
-  <groupId>org.opendaylight.lispflowmapping</groupId>
   <artifactId>mappingservice.implementation</artifactId>
-  <version>1.3.0-SNAPSHOT</version>
   <packaging>bundle</packaging>
   <name>Mapping Service Implementation</name>
 
@@ -21,15 +19,14 @@
       <version>${project.version}</version>
     </dependency>
     <dependency>
+        <groupId>${project.groupId}</groupId>
+        <artifactId>mappingservice.inmemorydb</artifactId>
+        <version>${project.version}</version>
+    </dependency>
+    <dependency>
       <groupId>${project.groupId}</groupId>
       <artifactId>mappingservice.yangmodel</artifactId>
       <version>${project.version}</version>
-    </dependency>
-
-    <!-- needed for activator -->
-    <dependency>
-      <groupId>org.apache.felix</groupId>
-      <artifactId>org.apache.felix.dependencymanager</artifactId>
     </dependency>
 
     <!-- needed for osgi console -->
