+++ mappingservice/inmemorydb/src/main/java/org/opendaylight/controller/config/yang/config/lfm/ms/dao/inmemorydb/InMemoryDbModuleFactory.java
@@ -1,0 +1,13 @@
+/*
+* Generated file
+*
+* Generated from: yang module name: lfm-ms-inmemorydb yang module local name: lfm-ms-inmemorydb
+* Generated by: org.opendaylight.controller.config.yangjmxgenerator.plugin.JMXGenerator
+* Generated at: Mon Jul 13 12:44:05 CEST 2015
+*
+* Do not modify this file unless it is present under src/main directory
+*/
+package org.opendaylight.controller.config.yang.config.lfm.ms.dao.inmemorydb;
+public class InMemoryDbModuleFactory extends org.opendaylight.controller.config.yang.config.lfm.ms.dao.inmemorydb.AbstractInMemoryDbModuleFactory {
+
+}
